+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.4.0
+
+* New image storadev/kubedoom:0.4.0
+* New `-mode` flag to switch between killing pods or namespaces.
+* Update kubernetes to 1.18.2
+
 # 0.3.0
 
 * New image storadev/kubedoom:0.3.0